+++ crates/scott/src/scottfree_save.rs
@@ -205,7 +205,6 @@ mod tests {
                 .collect(),
             adventure_number: 0,
             mysterious: false,
-            second_person: false,
             saga_us: None,
             ti99: None,
         }